#  perf: optimize buffer array allocations using np.asarray to avoid unnecessary copies