@araffin 2025-07-21
> Output: ~2-3x speedup even for the copy case

I hope you didn't use a LLM for that... when I run the code, I don't see any difference.

```
np.array(): 0.0188s
np.asarray().copy(): 0.0186s
Speedup: 1.0x
```

on two different machines.

> 30% improvement in typical RL training scenarios

What did you use to come up to that conclusion?